+++ src/main/java/janggi/domain/piece/Soldier.java
@@ -1,7 +1,7 @@
 package janggi.domain.piece;
 
 import janggi.domain.Piece;
-import janggi.domain.Position;
+import janggi.domain.position.Position;
 import janggi.domain.Team;
 import java.util.List;
 import java.util.Objects;
