+++ src/main/java/janggi/domain/position/Row.java
@@ -1,22 +1,22 @@
-package janggi.domain;
+package janggi.domain.position;
 
 import java.util.Objects;
 
-public class Column {
+public class Row {
 
-    public static final int COLUMN_LOWER_THRESH_HOLD = 1;
-    public static final int COLUMN_UPPER_THRESH_HOLD = 9;
+    public static final int ROW_LOWER_THRESH_HOLD = 1;
+    public static final int ROW_UPPER_THRESH_HOLD = 10;
 
     private final int value;
 
-    public Column(int value) {
+    public Row(int value) {
         validateRange(value);
         this.value = value;
     }
 
     private void validateRange(int value) {
-        if (value < COLUMN_LOWER_THRESH_HOLD || value > COLUMN_UPPER_THRESH_HOLD) {
-            throw new IllegalArgumentException("[ERROR] 열 좌표는 1~9까지 사용 가능 합니다");
+        if (value < ROW_LOWER_THRESH_HOLD || value > ROW_UPPER_THRESH_HOLD) {
+            throw new IllegalArgumentException("[ERROR] 행 좌표는 1~10까지 사용 가능 합니다");
         }
     }
 
@@ -32,8 +32,8 @@ public class Column {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Column column = (Column) o;
-        return value == column.value;
+        Row row = (Row) o;
+        return value == row.value;
     }
 
     @Override
