+++ src/test/java/janggi/domain/board/BoardTest.java
@@ -1,10 +1,11 @@
-package janggi.domain;
+package janggi.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import janggi.domain.piece.EmptyPiece;
+import janggi.domain.Piece;
+import janggi.domain.position.Position;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
