+++ src/test/java/janggi/domain/piece/GeneralTest.java
@@ -3,10 +3,10 @@ package janggi.domain.piece;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import janggi.domain.Column;
+import janggi.domain.position.Column;
 import janggi.domain.Piece;
-import janggi.domain.Position;
-import janggi.domain.Row;
+import janggi.domain.position.Position;
+import janggi.domain.position.Row;
 import janggi.domain.Team;
 import java.util.LinkedHashMap;
 import java.util.List;
